test(mcp): remove stdio startup from server tests
